+++ packages/bridge-controller/src/index.ts
@@ -94,7 +94,6 @@ export {
   SortOrder,
   ChainId,
   RequestStatus,
-  FeatureId,
   type TokenFeature,
   type QuoteStreamCompleteData,
   type BridgeControllerGetStateAction,
@@ -105,12 +104,15 @@ export {
   DiscountType,
   FeeType,
   ActionTypes,
-  BridgeAssetSchema,
-  TokenFeatureType,
+} from './validators/quote-response';
+export {
   validateQuoteStreamComplete,
   QuoteStreamCompleteReason,
-  BatchSellTransactionType,
-} from './utils/validators';
+} from './validators/quote-stream-complete';
+export { BatchSellTransactionType } from './validators/batch-sell';
+export { TokenFeatureType } from './validators/token-feature';
+export { BridgeAssetSchema } from './validators/bridge-asset';
+export { FeatureId } from './validators/feature-flags';
 
 export {
   ALLOWED_BRIDGE_CHAIN_IDS,
