+++ packages/bridge-controller/src/utils/metrics/properties.ts
@@ -9,7 +9,6 @@ import type {
   QuoteResponseV1,
   TxData,
 } from '../../types';
-import { FeatureId } from '../../types';
 import { getNativeAssetForChainId, isCrossChain } from '../bridge';
 import {
   formatAddressToAssetId,
@@ -23,6 +22,7 @@ import type {
   QuoteWarning,
   RequestParams,
 } from './types';
+import { FeatureId } from '../../validators/feature-flags';
 
 export const toInputChangedPropertyKey: Partial<
   Record<keyof QuoteRequest, InputKeys>
