+++ packages/bridge-controller/src/utils/quote.ts
@@ -17,7 +17,7 @@ import type {
   NonEvmFees,
   TxData,
 } from '../types';
-import { FeatureId } from '../types';
+import { FeatureId } from '../validators/feature-flags';
 import { isNativeAddress, isNonEvmChainId } from './bridge';
 
 export const isValidQuoteRequest = (
@@ -152,9 +152,7 @@ export const calcSentAmount = (
   const sentAmount = intent
     ? new BigNumber(srcTokenAmount)
     : Object.values(feeData)
-        .filter(
-          (fee) => fee && fee.amount && fee.asset?.assetId === srcAsset.assetId,
-        )
+        .filter((fee) => fee?.amount && fee.asset?.assetId === srcAsset.assetId)
         .reduce(
           (acc, { amount }) => acc.plus(amount),
           new BigNumber(srcTokenAmount),
