+++ packages/bridge-controller/src/validators/bridge-asset.ts
@@ -1,0 +1,50 @@
+import {
+  number,
+  type,
+  string,
+  optional,
+  nullable,
+  is,
+} from '@metamask/superstruct';
+import type { Infer } from '@metamask/superstruct';
+import { CaipAssetTypeStruct } from '@metamask/utils';
+
+export const ChainIdSchema = number();
+
+export const BridgeAssetSchema = type({
+  /**
+   * The chainId of the token
+   */
+  chainId: ChainIdSchema,
+  /**
+   * An address that the metaswap-api recognizes as the default token
+   */
+  address: string(),
+  /**
+   * The assetId of the token
+   */
+  assetId: CaipAssetTypeStruct,
+  /**
+   * The symbol of token object
+   */
+  symbol: string(),
+  /**
+   * The name for the network
+   */
+  name: string(),
+  decimals: number(),
+  /**
+   * URL for token icon
+   */
+  icon: optional(nullable(string())),
+  /**
+   * URL for token icon
+   */
+  iconUrl: optional(nullable(string())),
+});
+
+export const validateBridgeAsset = (
+  data: unknown,
+): data is Infer<typeof BridgeAssetSchema> => {
+  return is(data, BridgeAssetSchema);
+};
