+++ packages/bridge-controller/tests/mock-quotes-native-erc20-eth.ts
@@ -1,7 +1,10 @@
 import { merge } from 'lodash';
 
 import type { QuoteResponseV1, DeepPartial } from '../src/types';
-import { ActionTypes, validateQuoteResponseV1 } from '../src/utils/validators';
+import {
+  ActionTypes,
+  validateQuoteResponseV1,
+} from '../src/validators/quote-response';
 
 export const mockBridgeQuotesNativeErc20EthV1: QuoteResponseV1[] = [
   {
